+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Version
 
 ### Features Added
+- Migrate from the deprecated `vertx-web-api-contract` module to `vertx-web-openapi` [#506](https://github.com/ConsenSys/web3signer/pull/506)
 - Migrate jackson `ObjectMapper` instances to `JsonMapper` and `YamlMapper` builders to resolve deprecation warnings [#507](https://github.com/ConsenSys/web3signer/pull/507)
 
 ### Bugs Fixed